refactor spiBus: use dependency injection for the entire SPI object. Makes it way more configurable.
NOTE: the added flexibility might pose a difficulty obstacle for the user. Could maybe be mitigated by providing examples.
